% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Chromosome_data}
\alias{Chromosome_data}
\title{Chromosome Data}
\format{
A data frame with 2 columns and 4 rows
\describe{
\item{\code{No.of.Asso}}{No of Associations}
\item{\code{fre}}{Observed frequencies}
}
}
\source{
Extracted from

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in
Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: \doi{10.1080/03610928508828990}
}
\usage{
Chromosome_data
}
\description{
Data in this example refer to 337 observations on the secondary
association of chromosomes in Brassika; n , which is now the number of
chromosomes, equals 3 and X is the number of pairs of bivalents showing association.
}
\examples{
Chromosome_data$No.of.Asso          #extracting the binomial random variables
sum(Chromosome_data$fre)            #summing all the frequencies

}
\keyword{datasets}
