% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Course_data}
\alias{Course_data}
\title{Course Data}
\format{
A data frame with 2 columns and 9 rows
\describe{
\item{\code{sub.pass}}{subjects passed}
\item{\code{fre}}{Observed frequencies}
}
}
\source{
Extracted from

Karlis, D. & Xekalaki, E., 2008. The Polygonal Distribution. In Advances in Mathematical and Statistical
Modeling. Boston: Birkhuser Boston, pp. 21-33.

Available at: \doi{10.1007/978-0-8176-4626-4_2}.
}
\usage{
Course_data
}
\description{
The data refer to the numbers of courses taken by a class of 65 students from the first year of the
Department of Statistics of Athens University of Economics. The students enrolled in this class
attended 8 courses during the first year of their study. The total numbers of successful
examinations (including resits) were recorded.
}
\examples{
Course_data$sub.pass             # extracting the binomial random variables
sum(Course_data$fre)             # summing all the frequencies

}
\keyword{datasets}
