% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddBin.R
\name{fitAddBin}
\alias{fitAddBin}
\title{Fitting the Additive Binomial Distribution when binomial
random variable, frequency, probability of success and alpha are given}
\usage{
fitAddBin(x,obs.freq,p,alpha)
}
\arguments{
\item{x}{vector of binomial random variables.}

\item{obs.freq}{vector of frequencies.}

\item{p}{single value for probability of success.}

\item{alpha}{single value for alpha.}
}
\value{
The output of \code{fitAddBin} gives the class format \code{fitAB} and \code{fit} consisting a list

\code{bin.ran.var} binomial random variables.

\code{obs.freq} corresponding observed frequencies.

\code{exp.freq} corresponding expected frequencies.

\code{statistic} chi-squared test statistics.

\code{df} degree of freedom.

\code{p.value} probability value by chi-squared test statistic.

\code{fitAB} fitted probability values of \code{dAddBin}.

\code{NegLL} Negative Log Likelihood value.

\code{p} estimated probability value.

\code{alpha} estimated alpha parameter value.

\code{AIC} AIC value.

\code{call} the inputs of the function.

Methods \code{summary}, \code{print}, \code{AIC}, \code{residuals} and \code{fitted}
can be used to extract specific outputs.
}
\description{
The function will fit the Additive Binomial distribution when random variables,
corresponding frequencies, probability of success and alpha are given.
It will provide the expected frequencies, chi-squared test statistics value, p value,
and degree of freedom value so that it can be seen if this distribution fits the data.
}
\details{
\deqn{obs.freq \ge 0}
\deqn{x = 0,1,2,..}
\deqn{0 < p < 1}
\deqn{-1 < alpha < 1}
}
\examples{
No.D.D <- 0:7         #assigning the random variables
Obs.fre.1 <- c(47,54,43,40,40,41,39,95)            #assigning the corresponding the frequencies

\dontrun{
#assigning the estimated probability value
paddbin <- EstMLEAddBin(No.D.D,Obs.fre.1)$p

#assigning the estimated alpha value
alphaaddbin <- EstMLEAddBin(No.D.D,Obs.fre.1)$alpha

#fitting when the random variable,frequencies,probability and alpha are given
results <- fitAddBin(No.D.D,Obs.fre.1,paddbin,alphaaddbin)
results

#extracting the AIC value
AIC(results)

#extract fitted values
fitted(results)
       }

}
\references{
Johnson, N. L., Kemp, A. W., & Kotz, S. (2005). Univariate discrete distributions (Vol. 444).
Hoboken, NJ: Wiley-Interscience.

L. L. Kupper, J.K.H., 1978. The Use of a Correlated Binomial Model for the Analysis of Certain Toxicological
Experiments. Biometrics, 34(1), pp.69-76.

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics
- Theory and Methods, 14(6), pp.1497-1506.

Available at: \url{http://www.tandfonline.com/doi/abs/10.1080/03610928508828990} .

Jorge G. Morel and Nagaraj K. Neerchal. Overdispersion Models in SAS. SAS Institute, 2012.
}
