% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Terror_data_ARG}
\alias{Terror_data_ARG}
\title{Terror Data ARG}
\format{
A data frame with 2 columns and 9 rows
\describe{
\item{\code{Incidents}}{No of Incidents Occurred}
\item{\code{fre}}{Observed frequencies}
}
}
\source{
Extracted from

Li, X. H., Huang, Y. Y., & Zhao, X. Y. (2011). The Kumaraswamy Binomial Distribution. Chinese Journal
of Applied Probability and Statistics, 27(5), 511-521.
}
\usage{
Terror_data_ARG
}
\description{
Jenkins and Johnson (1975) compiled a chronology of incidents of international terrorism from 1/1968
through 04/1974. During this period 507 incidents are recorded in the world, where 64 incidents occurred
in the United States and 65 ones in Argentina.
}
\examples{
Terror_data_ARG$Incidents        #extracting the binomial random variables
sum(Terror_data_ARG$fre)              #summing all the frequencies

}
\keyword{datasets}
