% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Alcohol_data}
\alias{Alcohol_data}
\title{Alcohol data}
\format{
A data frame with 3 columns and 8 rows.
\describe{
\item{\code{Days}}{No of Days Drunk}
\item{\code{week1}}{Observed frequencies for week1}
\item{\code{week2}}{Observed frequencies for week2}
}
}
\source{
Extracted from

Manoj, C., Wijekoon, P. & Yapa, R.D., 2013. The McDonald Generalized Beta-Binomial Distribution: A New
Binomial Mixture Distribution and Simulation Based Comparison with Its Nested Distributions in Handling
Overdispersion. International Journal of Statistics and Probability, 2(2), pp.24-41.

Available at: \doi{10.5539/ijsp.v2n2p24}
}
\usage{
Alcohol_data
}
\description{
Lemmens , Knibbe and Tan(1988) described a study of self reported alcohol frequencies.
The no of alcohol consumption data in two reference weeks is separately self reported
by a randomly selected sample of 399 respondents in the Netherlands in 1983.
Number of days a  given individual consumes alcohol out of 7 days a week can be treated as a binomial
variable.
The collection of all such variables from all respondents would be defined as "Binomial Outcome Data".
}
\examples{
Alcohol_data$Days          # extracting the binomial random variables
sum(Alcohol_data$week2)       # summing all the frequencies in week2

}
\keyword{datasets}
